fix: Jest encountered an unexpected token
yarn test was giving an error which said 
"Jest encountered an unexpected token"
I was able to solve it by removing transformIgnorePatterns 
from jest config in package.json.

To be exact, transformIgnorePatterns didn't include 
unimodules-permissions-interface
and that needed to transformed for jest to compile files successfully.
For now, I will comment out the transformIgnorePatterns.
If this is causing some error, further investigation will be made.

Additionally, I created jest.config.js in order for better organisation.