fix: Fix sounds not working
I seperated the timer files that are of different speeds.
This ensures smooth shifting of speed of clock ticking, 
as well as better handling of async calls.

I will still have to handle cases when the user tries to go back to 
previous
but will also disable going back in the next commit.