fix: Prevent the game from getting stuck
The game was getting stuck in Android. 
This is most likely due to the fact that the music files were not being 
loaded/unloaded properly. 
This commit removes the Unhandled promise rejection error which says 
Cannot complete operation because sound is not loaded.

As you can see, this commit mearly resolved the error. 
Refactoring is needed to make the code more readable.

Related: #8